CORRECTED: Clarify error message in scripts/lambdanative
The current source checks two different error conditions in one command and uses the same uninformative error.  This patch separates the tests, clarifies the error, and tells the user how to fix the problem.

Tested on OSX but uses nothing but sh builtins so should be cross platform.

NOTE:  Accidentally clicked 'submit' on an erroneous version of this that mistakenly chopped the "X$PWD" test.  Please ignore that one.